docs: htlc_accepted cltv_expiry_relative is signed
lightningd emits this field with json_add_s32 (the difference
between the HTLC's cltv_expiry and the current blockheight can
be negative):

  lightningd/peer_htlcs.c:1200
    s32 expiry = hin->cltv_expiry, blockheight = ...
  lightningd/peer_htlcs.c:1252
    json_add_s32(s, "cltv_expiry_relative", expiry - blockheight);

Type it as `integer` rather than `u32`.

Changelog-None

+++ cln-rpc/src/hooks.rs
@@ -242,7 +242,7 @@ pub mod events{
         pub extra_tlvs: Option<String>,
         pub amount_msat: Amount,
         pub cltv_expiry: u32,
-        pub cltv_expiry_relative: u32,
+        pub cltv_expiry_relative: i64,
         pub id: u64,
         pub payment_hash: Sha256,
         pub short_channel_id: ShortChannelId,
